pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/llm"]

[project]
name = "assetopsbench-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "INSTRUCTIONS.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2024.8.30",
    "couchdb3>=2.0.2",
    "fastmcp>=2.14.5",
    "mcp[cli]>=1.26.0",
    "pendulum>=3.2.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "litellm==1.81.13",
    "claude-agent-sdk>=0.0.14",
    "python-dotenv>=1.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.8.0",
    "datasets>=4.8.4",
    "pypdf>=6.9.2",
    "torch>=2.0",
    "wandb>=0.18",
]

[project.scripts]
plan-execute = "agent.cli:main"
claude-agent = "agent.claude_agent.cli:main"
iot-mcp-server = "servers.iot.main:main"
utilities-mcp-server = "servers.utilities.main:main"
fmsr-mcp-server = "servers.fmsr.main:main"
tsfm-mcp-server = "servers.tsfm.main:main"
wo-mcp-server = "servers.wo.main:main"
vibration-mcp-server = "servers.vibration.main:main"


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "anyio>=4.9.0",
    "pytest-anyio>=0.0.0",
]
# Optional heavy ML deps for the TSFM server.
# tsfm_public must be installed separately: pip install git+https://github.com/ibm-granite/granite-tsfm
tsfm = [
    "torch>=2.0",
    "transformers>=4.40",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
norecursedirs = ["src/tmp"]
filterwarnings = [
    "ignore:Core Pydantic V1 functionality:UserWarning",
]

